:truck: Move Spell Card model to cards/src/
