+++ source/YugiAcademy/YuGiOh/models.py
@@ -4,13 +4,6 @@ from django.contrib.contenttypes.fields import GenericForeignKey, GenericRelatio
 from django.contrib.contenttypes.models import ContentType
 from django.db import models
 
-from utils.src.instance_creation_failed import InstanceCreationFailed
-
-
-def enforce_not_blank(string, name):
-    if not string.strip():
-        raise InstanceCreationFailed(f'{name} must not be blank.')
-
 
 class BoosterPack(models.Model):
     name = models.CharField(max_length=50, unique=True)
@@ -65,21 +58,6 @@ class MonsterCard(Card):
 
     def related_query_name(self):
         return 'monster_card'
-
-
-class SpellCard(Card):
-    type = models.CharField(max_length=20)
-
-    @classmethod
-    def named(cls, name: str, type: str, description: str):
-        enforce_not_blank(name, "Name")
-        enforce_not_blank(type, "Type")
-        enforce_not_blank(description, "Description")
-
-        return cls(name=name, type=type, description=description)
-
-    def related_query_name(self):
-        return 'spell_card'
 
 
 class TrapCard(Card):
